* Expose a client_throttle_multiplier setting in OpenSim.ini.  This multiplier is applied to all the client throttle settings received by the client * This should probably be 1, but currently by default it is 8, to reflect what was being eon3 in OpenSim before this revision.  So if the client requested a maximum throttle of 1500 kilobits per second, we would actually send out 1500 kilobytes per second * Adjusting this multiplier down towards 1 may improve your OpenSim experience, though in other situations it may degrade (e.g. if you're using a standalone over high bandwidth links) * This is currently a user setting because adjusting it down may currently reveal other OpenSim bugs.

+++ OpenSim/Region/ClientStack/ClientStackUserSettings.cs
@@ -44,6 +44,6 @@ namespace OpenSim.Region.ClientStack
         /// A multiplier applied to all client throttle settings.  This is hopefully a temporary setting to iron out
         /// bugs that appear if the existing incorrect * 8 throttle (bytes instead of bits) is corrected.
         /// </summary>
-        public int ClientThrottleMultipler;
+        public float ClientThrottleMultipler;
     }
 }
